backend/requirements.txt
# FastAPI & Server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
httpx==0.26.0

# Auth & Security
python-jose[cryptography]==3.3.0
pyjwt==2.8.0
cryptography==42.0.0

# Auth0 AI SDK
auth0-ai==0.1.0

# Salesforce
simple-salesforce==1.12.5

# Azure AI Foundry
azure-identity==1.15.0
azure-ai-inference==1.0.0b1
openai==1.12.0

# Data & Validation
pydantic==2.6.0
pydantic-settings==2.1.0

# Database (for demo data)
sqlalchemy==2.0.25
aiosqlite==0.19.0

# Utils
python-dotenv==1.0.1
tenacity==8.2.3
structlog==24.1.0
